ChessBrowser: PgnParser and unit tests
Implements a PgnParser class to begin addressing T239438. The
parser currently only parses simple PGN documents. It still fails
on variations because the JavaScript module does not support
variations. The output is legal PGN which the JS module is able
to parse. To resolve T239438, the JS module needs modification
to take input other than a PGN document so that it does not
need its own parser, and all parsing can be done server-side.

Addresses T239529 by implementing PHPunit tests for the
PgnParser class. Most methods are explicitly tested. The
ChessBrowser class still requires tests to be written, and
some edge cases and fine points of the PgnParser class still
need tests written for them.

The JavaScript will be rewritten in the near future (see T240742)
and so technical debt is incurred to prevent wasting effort on
code soon to be rewritten. Linting errors which have been ignored
for that reason have been marked as having incurred technical debt.

Bug: T239438, T239529
Change-Id: Id2f049fcee033d7eacb665217e0b97c2aa1952ca