Submodule and license: Add PGN Parser Submodule
Add a fork of the DHTMLchess PgnParser as a subtree. Because that
program is licensed under the LGPL, the combined work must be licensed
under a version of the (L)GPL. Therefore this commit changes the license
of the ChessBrowser program to the General Public License v3 or later.

Adds .eslintignore and .stylelintignore files to prevent build errors.

Bug: T239438
Change-Id: Ica2dfe578aef7d0a2f35ada03cd8bdd62f7f64d9