javascript: implement new JS module code
Add and modify the new JavaScript from Kipod and integrate it into
the extension. The module takes the PHP's JSON output and
provides the interactive viewer on the client side.

The PHP is modified to keep track of how many boards are on the
page and pass a list of identifiers as a config variable
so that the module can handle arbitrarily many games on a page
without conflict.

Bug: T240742
Change-Id: Iff98fd9af423bc7225a75196b27a401c8dd4e07b