Remove auto-play feature
The auto-play feature complicates the extension's user interface.

* It makes the user interface harder to understand. The meaning of
  'walking person' and 'bicycle' is not obvious. I am not sure that
  there is a suitable replacement icon that would convey how these
  buttons behave. I think this makes the extension less inviting to new
  users.
* It makes it harder to provide a good, consistent experience across a
  broad range of devices. The feature adds three buttons to the toolbar,
  bringing the total to eight. The recommended minimum size for
  touch-interface buttons is around 11mm[1], so 8 buttons = 88mm (with
  no spacing), which is already wider than many mobile phones.
* JavaScript code that uses timers is bug-prone. I don't have data for
  this, but that has been my experience.

The feature could always be added back if users clamor for it. For the
initial launch, I think it's best to pare down features to the
essentials. Features can become obstacles to further development and
refinement of this extension because it is harder to take features away
than it is to add them.

Please see the associated task (T281649) for before/after screenshots.

  [1]: https://ux.stackexchange.com/questions/39023/what-is-the-optimum-button-size-of-touch-screen-applications

Bug: T281649
Change-Id: If171cb4c51b3cfd1bacd8621ea54fd0e7a7e25a8