Implement comment and variation parsing
Modify the regex to allow commented PGN to pass validation.

Comments are now passed as part of the ply information in the
"special" field following the castling information and such.

Variations are passed to the JS as the data attribute of
a span. The data attribute contains all variations together
at the top level. The data object is an array of
arrays where each sub array is the move information for
the variation.

Add and modify templates for variation and comment generation
and styling.

Bug: T284034
Change-Id: I299c5d798546aaec6d6bdc9653eb7a9adda66c0f