Add XML args for starting state and perspective
The starting board state can be specified using "ply=N" where
"N" is the ply number to display. If not specified, it defaults
to the first move of the game.

The board perspective can be set using "side='black'" to load
the FEN or game viewer from black's perspective. Defaults to
white's perspective. This uses the pgn-flip CSS class, so some
tweaks were made to suppress the board-spinning animation on
load but restore it for user input.

Bug: T290285
Change-Id: I4f1cdb8426bb10fbd44aa1811f4b0e7a33d41d77