Context sensitive comment parsing and EOL comments
PGN allows end-of-line comments with a semicolon denoting that the rest
of the following line is a comment. Previously, the program ignored
these types of comments. This patch handles them and makes them visible
to the reader.

To do so properly, upstream PgnParser is modified to implement
context-sensitive parsing of comments to determine whether a character
has its special or normal meaning. Tests are included for these corner
cases.

Bug: T291952, T363230
Change-Id: Idca9dded9e3f9094f376eee40b7be17a9a400569